+++ CHANGES.rst
@@ -18,6 +18,7 @@ API changes
 - Overhauled the mock stellar stream generation methodology to allow for more
   general stream generation. See ``MockStreamGenerator`` and the stream
   distribution function classes, e.g., ``FardalStreamDF``.
+- Removed deprecated ``CartesianPhaseSpacePosition`` class.
 
 
 1.0 (2019-04-12)